moving types to lib
